BREAKING(semver): make invalid semver constants private
