+++ semver/is_semver_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { assert } from "@std/assert";
-import { MAX, MIN } from "./constants.ts";
+import { MAX, MIN } from "./_constants.ts";
 import { isSemVer } from "./is_semver.ts";
 
 Deno.test({
