+++ semver/max_satisfying_test.ts
@@ -4,7 +4,7 @@ import { assertEquals } from "@std/assert";
 import { parse } from "./parse.ts";
 import { parseRange } from "./parse_range.ts";
 import { maxSatisfying } from "./max_satisfying.ts";
-import { MAX, MIN } from "./constants.ts";
+import { MAX, MIN } from "./_constants.ts";
 
 Deno.test({
   name: "maxSatisfying()",
